+++ apis/cassandra/v1beta1/webhook_test.go
@@ -247,6 +247,25 @@ func Test_ValidateSingleDatacenter(t *testing.T) {
 			},
 			errString: "use multiple nodes per worker without cpu and memory requests and limits",
 		},
+		{
+			name: "Prevent user specified reserved Service labels and annotations",
+			dc: &CassandraDatacenter{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "exampleDC",
+				},
+				Spec: CassandraDatacenterSpec{
+					ServerType:    "cassandra",
+					ServerVersion: "4.0.4",
+					AdditionalServiceConfig: ServiceConfig{
+						DatacenterService: ServiceConfigAdditions{
+							Labels:      map[string]string{"k8ssandra.io/key1": "val1", "cassandra.datastax.com/key2": "val2"},
+							Annotations: map[string]string{"k8ssandra.io/key3": "val3", "cassandra.datastax.com/key4": "val4"},
+						},
+					},
+				},
+			},
+			errString: "configure DatacenterService with reserved annotations and/or labels (prefixes cassandra.datastax.com and/or k8ssandra.io)",
+		},
 	}
 
 	for _, tt := range tests {