+++ pkg/reconciliation/construct_podtemplatespec.go
@@ -6,6 +6,7 @@ package reconciliation
 // This file defines constructors for k8s objects
 
 import (
+	"encoding/json"
 	"fmt"
 	"reflect"
 	"sort"
@@ -13,6 +14,7 @@ import (
 	"github.com/pkg/errors"
 
 	api "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
+	"github.com/k8ssandra/cass-operator/pkg/cdc"
 	"github.com/k8ssandra/cass-operator/pkg/httphelper"
 	"github.com/k8ssandra/cass-operator/pkg/images"
 	"github.com/k8ssandra/cass-operator/pkg/oplabels"
@@ -103,16 +105,21 @@ func selectorFromFieldPath(fieldPath string) *corev1.EnvVarSource {
 	}
 }
 
-func probe(port int, path string, initDelay int, period int) *corev1.Probe {
+func httpGetAction(port int, path string) *corev1.HTTPGetAction {
+	return &corev1.HTTPGetAction{
+		Port: intstr.FromInt(port),
+		Path: path,
+	}
+}
+
+func probe(port int, path string, initDelay int, period int, timeout int) *corev1.Probe {
 	return &corev1.Probe{
 		ProbeHandler: corev1.ProbeHandler{
-			HTTPGet: &corev1.HTTPGetAction{
-				Port: intstr.FromInt(port),
-				Path: path,
-			},
+			HTTPGet: httpGetAction(port, path),
 		},
 		InitialDelaySeconds: int32(initDelay),
 		PeriodSeconds:       int32(period),
+		TimeoutSeconds:      int32(timeout),
 	}
 }
 
@@ -358,11 +365,14 @@ func getConfigDataEnVars(dc *api.CassandraDatacenter) ([]corev1.EnvVar, error) {
 	}
 
 	configData, err := dc.GetConfigAsJSON(dc.Spec.Config)
-
 	if err != nil {
 		return envVars, err
 	}
-	envVars = append(envVars, corev1.EnvVar{Name: "CONFIG_FILE_DATA", Value: configData})
+	cdcAdded, err := cdc.UpdateConfig(json.RawMessage(configData), *dc)
+	if err != nil {
+		return envVars, err
+	}
+	envVars = append(envVars, corev1.EnvVar{Name: "CONFIG_FILE_DATA", Value: string(cdcAdded)})
 
 	return envVars, nil
 }
@@ -421,11 +431,11 @@ func buildContainers(dc *api.CassandraDatacenter, baseTemplate *corev1.PodTempla
 	}
 
 	if cassContainer.LivenessProbe == nil {
-		cassContainer.LivenessProbe = probe(8080, "/api/v0/probes/liveness", 15, 15)
+		cassContainer.LivenessProbe = probe(8080, httphelper.LivenessEndpoint, 15, 15, 10)
 	}
 
 	if cassContainer.ReadinessProbe == nil {
-		cassContainer.ReadinessProbe = probe(8080, "/api/v0/probes/readiness", 20, 10)
+		cassContainer.ReadinessProbe = probe(8080, httphelper.ReadinessEndpoint, 20, 10, 10)
 	}
 
 	if cassContainer.Lifecycle == nil {
@@ -433,7 +443,7 @@ func buildContainers(dc *api.CassandraDatacenter, baseTemplate *corev1.PodTempla
 	}
 
 	if cassContainer.Lifecycle.PreStop == nil {
-		action, err := httphelper.GetMgmtApiWgetPostAction(dc, httphelper.WgetNodeDrainEndpoint, "")
+		action, err := httphelper.GetMgmtApiWgetPostAction(dc, httphelper.NodeDrainEndpoint, "", 0)
 		if err != nil {
 			return err
 		}